Moved demos to ts src
