--- packages/io-markdown/demos/IoMarkdownDemo.js
@@ -1,7 +1,0 @@
-import { Register, IoElement } from 'io-gui';
-
-export class IoMarkdownDemo extends IoElement {
-  ready() {}
-}
-Register(IoMarkdownDemo);
-export const ioMarkdownDemo = IoMarkdownDemo.vConstructor;
