+++ packages/io-markdown/src/demos/IoMarkdownDemo.ts
@@ -1,0 +1,8 @@
+//@ts-nocheck
+import { Register, IoElement } from 'io-gui';
+
+export class IoMarkdownDemo extends IoElement {
+  ready() {}
+}
+Register(IoMarkdownDemo);
+export const ioMarkdownDemo = IoMarkdownDemo.vConstructor;
