+++ packages/io-sliders/rollup.config.js
@@ -2,5 +2,6 @@ import { makeBundleTarget } from '../../rollup.config.js';
 
 export default [
   makeBundleTarget('packages/io-sliders/build/index.js', 'packages/io-sliders/bundle/index.js'),
+  makeBundleTarget('packages/io-sliders/build/demos/IoSlidersDemo.js', 'packages/io-sliders/bundle/demos/IoSlidersDemo.js'),
   makeBundleTarget('packages/io-sliders/build/index.test.js', 'packages/io-sliders/bundle/index.test.js'),
 ];
